test: `Lotto` 테스트 이동
